moved a function to utils
